Change existingUser check
Testing of createUser requires knowing how it works internally. This
is due to existingUser being checked against null when the findUnique
mock call is not called, resulting in existingUser being undefined.

Let's change this check to be if (existingUser) so that both null
and undefined are treated as falsy. This removes the need to mock
the findUnique call in every test for createUser.