Change status code for POST with user already in database
The 400 error when POSTing to the database when the user is already
inside can surprise developers during debugging because everytime a
user logs in, the POST is made regardless of whether the user is inside
the database or not.

Let's change the status code to 200 to avoid showing it as an error
while still differentiating it from the status code 201 when the user
details are added for the first time.