Standardise naming and dependencies for files and scripts
The yarn dev command behaves inconsistently across workspaces. In some
workspaces, dev reads in a .env file while in others, dev does not
read in the .env file. This leads to confusion as a developer may run
dev expecting the .env file to be read when the correct command is
dev:local.

Additionally, either nodemon or ts-node-dev, which both restart the
backend service upon file change, are used in the workspaces. Having
both at the same time increases the number of dependencies we need to
manage.

Let's standardise the usage of dev:local for reading the .env file, use
ts-node-dev across all the backend services and standardise the
swagger-related file names.