Change Docker Compose to use production-style images (#196)
Docker Compose has largely been sidelined as a tool for developing the
app since a script that does not use Docker for starting up all the 
services at once has been created.

Let's repurpose Docker Compose to be a prod checker. The same images 
that are built for Google Kubernetes Engine will also be built for
Docker Compose so that we can have a good idea of whether the app
will still work in production after it has already been proven to work
in development.