Allow non-admins to try and fail at adding/editing questions
One of the assignment 3 requirements is to show HTTP error messages
for unauthorized adding/editing of questions.

To make it easier to test the scenario of a user who is not an admin
trying to make the changes, let's allow these non-admins to see and
submit the forms for adding/editing questions. However, submitting
these forms should fail with the question list remaining unchanged
on the backend.